feat(AAiT-backend-2B):Add Validators and CommonResoponse

+++ aait/backend/group-2B/Application/Features/Notifications/Handlers/Commands/DeleteNotificationCommandHandler.cs
@@ -1,10 +1,11 @@
 using MediatR;
+using SocialSync.Application.Common.Responses;
 using SocialSync.Application.Contracts.Persistence;
 using SocialSync.Application.Features.Notifications.Requests.Commands;
 
 namespace SocialSync.Application.Features.Notifications.Handlers.Commands;
 
-public class DeleteNotificationCommandHandler : IRequestHandler<DeleteNotificationCommand, Unit>
+public class DeleteNotificationCommandHandler : IRequestHandler<DeleteNotificationCommand, CommonResponse<Unit>>
 {
         private readonly IUnitOfWork _unitOfWork;
 
@@ -13,12 +14,21 @@ public class DeleteNotificationCommandHandler : IRequestHandler<DeleteNotificati
             _unitOfWork = unitOfWork;
         }
 
-        public async Task<Unit> Handle(DeleteNotificationCommand request, CancellationToken cancellationToken)
+        public async Task<CommonResponse<Unit>> Handle(DeleteNotificationCommand request, CancellationToken cancellationToken)
         {
-            // add any necessary validation here before deleting the notification
+
             var notification = await _unitOfWork.NotificationRepository.GetAsync(request.NotificationId);
+            if(notification == null){
+                return CommonResponse<Unit>.Failure("Notification Not Found");
+            }
             await _unitOfWork.NotificationRepository.DeleteAsync(notification);
 
-            return Unit.Value;
+            if( await _unitOfWork.SaveAsync() == 0)
+            {
+                return CommonResponse<Unit>.Failure("Error While Saving Changes!");
+            }
+
+            return CommonResponse<Unit>.Success(Unit.Value);
+
         }
 }
